Add dedicated signup controller
