+++ backend-java/src/main/java/com/example/demo/SignupController.java
@@ -6,21 +6,19 @@ import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class AuthController {
+public class SignupController {
 
     record SignupRequest(String fullName, String email, String password) {}
-    record LoginRequest(String email, String password) {}
-
 
     private final UserRepository userRepository;
     private final EncryptionService encryptionService;
 
-    public AuthController(UserRepository userRepository, EncryptionService encryptionService) {
+    public SignupController(UserRepository userRepository, EncryptionService encryptionService) {
         this.userRepository = userRepository;
         this.encryptionService = encryptionService;
     }
@@ -37,7 +35,7 @@ public class AuthController {
         }
         UserEntity entity = new UserEntity();
         entity.setFullName(request.fullName());
-        entity.setEmail(encryptionService.encrypt(request.email()));
+        entity.setEmail(encryptedEmail);
         entity.setPassword(encryptionService.encrypt(request.password()));
         entity.setCreatedDate(Instant.now());
         entity.setLastLoggedInDate(Instant.now());
@@ -45,26 +43,5 @@ public class AuthController {
         userRepository.save(entity);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(Map.of("message", "Welcome, " + request.fullName() + "!"));
-    }
-
-    @PostMapping("/api/login")
-    public ResponseEntity<?> login(@RequestBody LoginRequest request) {
-        if (request.email() == null || request.password() == null) {
-            return ResponseEntity.badRequest().body(Map.of("error", "Email and password required"));
-        }
-        String encryptedEmail = encryptionService.encrypt(request.email());
-        Optional<UserEntity> optional = userRepository.findByEmail(encryptedEmail);
-        if (optional.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("error", "Invalid credentials"));
-        }
-        UserEntity user = optional.get();
-        String decryptedPassword = encryptionService.decrypt(user.getPassword());
-        if (!decryptedPassword.equals(request.password())) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(Map.of("error", "Invalid credentials"));
-        }
-        Instant previousLogin = user.getLastLoggedInDate();
-        user.setLastLoggedInDate(Instant.now());
-        userRepository.save(user);
-        return ResponseEntity.ok(Map.of("lastLoggedInDate", previousLogin != null ? previousLogin.toString() : null));
     }
 }
